Исправляет название массивов
