+++ src/const.js
@@ -1,8 +1,8 @@
-export const Colors = [`black`, `yellow`, `blue`, `green`, `pink`];
+export const colors = [`black`, `yellow`, `blue`, `green`, `pink`];
 
-export const Days = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
+export const days = [`mo`, `tu`, `we`, `th`, `fr`, `sa`, `su`];
 
-export const MonthNames = [
+export const monthNames = [
   `January`,
   `February`,
   `March`,
