+++ src/mock/task.js
@@ -1,12 +1,14 @@
-import {Colors} from '../const.js';
+import {colors} from '../const.js';
+import {getRandomArrayItem} from '../utils.js';
+import {getRandomDate} from '../utils.js';
 
-const DescriptionItems = [
+const descriptionItems = [
   `Изучить теорию`,
   `Сделать домашку`,
   `Пройти интенсив на соточку`,
 ];
 
-const DefaultRepeatingDays = {
+const defaultRepeatingDays = {
   'mo': false,
   'tu': false,
   'we': false,
@@ -16,7 +18,7 @@ const DefaultRepeatingDays = {
   'su': false,
 };
 
-const Tags = [
+const tags = [
   `homework`,
   `theory`,
   `practice`,
@@ -24,28 +26,8 @@ const Tags = [
   `keks`
 ];
 
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomIntegerNumber(0, array.length);
-
-  return array[randomIndex];
-};
-
-const getRandomIntegerNumber = (min, max) => {
-  return min + Math.floor(max * Math.random());
-};
-
-const getRandomDate = () => {
-  const targetDate = new Date();
-  const sign = Math.random() > 0.5 ? 1 : -1;
-  const diffValue = sign * getRandomIntegerNumber(0, 7);
-
-  targetDate.setDate(targetDate.getDate() + diffValue);
-
-  return targetDate;
-};
-
 const generateRepeatingDays = () => {
-  return Object.assign({}, DefaultRepeatingDays, {
+  return Object.assign({}, defaultRepeatingDays, {
     'mo': Math.random() > 0.5,
   });
 };
@@ -60,11 +42,11 @@ const generateTask = () => {
   const dueDate = Math.random() > 0.5 ? null : getRandomDate();
 
   return {
-    description: getRandomArrayItem(DescriptionItems),
+    description: getRandomArrayItem(descriptionItems),
     dueDate,
-    repeatingDays: dueDate ? DefaultRepeatingDays : generateRepeatingDays(),
-    tags: new Set(generateTags(Tags)),
-    color: getRandomArrayItem(Colors),
+    repeatingDays: dueDate ? defaultRepeatingDays : generateRepeatingDays(),
+    tags: new Set(generateTags(tags)),
+    color: getRandomArrayItem(colors),
     isFavorite: Math.random() > 0.5,
     isArchive: Math.random() > 0.5,
   };
